requirement.txt
-i https://pypi.org/simple
absl-py==0.7.1
astor==0.7.1
backcall==0.2.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
cloudpickle==0.8.0
configparser==5.0.0
coremltools==3.4
cycler==0.10.0
dask[array]==1.1.4
decorator==4.3.2
enum34==1.1.6
funcsigs==1.0.2
gast==0.2.2
git+https://github.com/akinoriosamura/tf-coreml@1d9caaa4c42019587e4c5e0d1ce543e6be8274f1#egg=tfcoreml
grpcio==1.19.0
h5py==2.9.0
html5lib==0.9999999
imageio==2.9.0
imgaug==0.4.0
ipdb==0.13.3
ipython-genutils==0.2.0
ipython==7.16.1
jedi==0.17.2
joblib==0.16.0
keras-applications==1.0.7
keras-preprocessing==1.0.9
kiwisolver==1.0.1
markdown==3.1
matplotlib==2.2.4
mock==2.0.0
networkx==2.2
numpy==1.16.2
opencv-python==4.0.0.21
parso==0.7.0
pbr==5.1.3
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pillow==5.4.1
prompt-toolkit==3.0.5
protobuf==3.7.0
ptyprocess==0.6.0
pydot==1.4.1
pygments==2.6.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
pywavelets==1.0.2
scikit-image==0.14.2
scikit-learn==0.23.1
scipy==1.2.1
shapely==1.7.0
six==1.12.0
subprocess32==3.5.3
termcolor==1.1.0
tensorflow_addons
threadpoolctl==2.1.0
toolz==0.9.0
torch==1.0.0
traitlets==4.3.3
wcwidth==0.2.5
werkzeug==0.15.1
